% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_scores.r
\name{collapse_score}
\alias{collapse_score}
\title{Collapse counts matrix to vector}
\usage{
collapse_score(counts, method)
}
\arguments{
\item{counts}{Matrix or numeric vector of counts.}

\item{method}{Character string containing "mean" or "sum" to specify
aggregation method.}
}
\value{
Numeric vector of aggregated values.
}
\description{
Collapses a counts matrix to a single vector using mean or sum aggregation.
}
