% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_scores.r
\name{compute_score}
\alias{compute_score}
\title{Compute score from two count matrices}
\usage{
compute_score(counts_1, counts_2, method)
}
\arguments{
\item{counts_1}{Matrix or vector representing counts for first gene set.}

\item{counts_2}{Matrix or vector representing counts for second gene set.}

\item{method}{Character string specifying "ratio" or "sub"
(subtraction) method.}
}
\value{
Numeric vector of computed scores.
}
\description{
Computes ratio or subtraction scores between two count matrices.
}
