% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coordinates}
\alias{coordinates}
\alias{coord}
\title{Spatial coordinates}
\format{
coordinate data frame
\describe{
 \item{coordinates}{spatial coordinates of a 2 cell circle synthetic
   regime produced oneiric}
}
}
\source{
produced by \url{https://github.com/WonLab-CS/oneiric}
}
\usage{
data(Kuresi)
}
\description{
Spatial coordinates
}
\keyword{datasets}
