% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.r
\name{create_palette}
\alias{create_palette}
\title{territories present. If required the colours will be randomly assinged
to each territory. Note that as the territory plot 
return a ggplot object, you can easily override the color scheme.}
\usage{
create_palette(territories, randomise)
}
\value{
color vector
}
\description{
territories present. If required the colours will be randomly assinged
to each territory. Note that as the territory plot 
return a ggplot object, you can easily override the color scheme.
}
