% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{function_call}
\alias{function_call}
\title{Generate function based on score type}
\usage{
function_call(score_type, single = TRUE, by_territory = TRUE)
}
\arguments{
\item{score_type}{Character string containing "mean" or "sum" to
specify function type.}

\item{single}{Logical whether working with single values
(default: TRUE).}

\item{by_territory}{Logical whether aggregating by territory
(default: TRUE).}
}
\value{
A function (mean, sum, Matrix::colMeans, or Matrix::colSums).
}
\description{
Returns an appropriate aggregation function (mean, sum, or Matrix
functions) based on score type and context parameters.
}
