% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.r
\name{generate_palette}
\alias{generate_palette}
\title{Create color palette to vizualize seed and query scores}
\usage{
generate_palette(palette, seed, query)
}
\description{
Create color palette to vizualize seed and query scores
}
