% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{load_visium}
\alias{load_visium}
\title{Loading Visium data anc creating a Vesalius object}
\usage{
load_visium(input_dir, tag = NULL)
}
\arguments{
\item{input_dir}{character string - path to directory containing visium data}

\item{tag}{character string - file tag used to define data set name}
}
\value{
vesalius_asssay object
}
\description{
Loading Visium data anc creating a Vesalius object
}
