% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_score.r
\name{outcomes_as_score_matrix}
\alias{outcomes_as_score_matrix}
\title{Convert outcomes to score matrix}
\usage{
outcomes_as_score_matrix(scores)
}
\arguments{
\item{scores}{A nested list containing competition outcomes from
compute_competition_outcomes.}
}
\value{
A matrix with average scores between all group pairs.
}
\description{
Converts nested list of competition outcomes into a matrix where each
entry represents the average score between groups.
}
