% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{win_lose_genes}
\alias{win_lose_genes}
\title{Create Win/Lose gene list}
\usage{
win_lose_genes()
}
\value{
A list containing two character vectors: "win"
  (pro-competitive genes) and "lose" (anti-competitive genes).
}
\description{
Returns predefined lists of "win" and "lose" genes associated with
competitive fitness in cancer cells.
}
